+++ src/dockerContainers.ts
@@ -9,6 +9,7 @@ export class DockerContainers implements vscode.TreeDataProvider<DockerContainer
     public _onDidChangeTreeData: vscode.EventEmitter<DockerContainer | undefined> = new vscode.EventEmitter<DockerContainer | undefined>();
     public readonly onDidChangeTreeData: vscode.Event<DockerContainer | undefined> = this._onDidChangeTreeData.event;
     private isErrorMessageShown = false;
+    private containerStrings = [];
 
     constructor(private context: vscode.ExtensionContext) {
         this.setAutoRefresh();
@@ -19,6 +20,26 @@ export class DockerContainers implements vscode.TreeDataProvider<DockerContainer
         this._onDidChangeTreeData.fire();
     }
 
+    public searchContainer(): void {
+        const interval = Utility.getConfiguration().get<number>("autoRefreshInterval");
+        let containerStrings = [];
+        if (interval > 0 && this.containerStrings.length > 0) {
+            this.containerStrings.forEach((containerString) => {
+                const items = containerString.split(" ");
+                containerStrings.push(`${items[1]} (${items[2]})`);
+            });
+        } else {
+            containerStrings = Executor.execSync("docker ps -a --format \"{{.Names}} ({{.Image}})\"").split(/[\r\n]+/g).filter((item) => item);
+        }
+
+        vscode.window.showQuickPick(containerStrings, { placeHolder: "Search Docker Container" }).then((containerString) => {
+            if (containerString !== undefined) {
+                const items = containerString.split(" ");
+                this.getContainer(items[0]);
+            }
+        });
+    }
+
     public getTreeItem(element: DockerContainer): vscode.TreeItem {
         return element;
     }
@@ -26,23 +47,22 @@ export class DockerContainers implements vscode.TreeDataProvider<DockerContainer
     public getChildren(element?: DockerContainer): Thenable<DockerContainer[]> {
         const containers = [];
         try {
-            const containerStrings = Executor.execSync("docker ps -a --format \"{{.ID}} {{.Names}} {{.Image}} {{.Status}}\"").split(/[\r\n]+/g);
-            containerStrings.forEach((containerString) => {
-                if (containerString) {
-                    const items = containerString.split(" ");
-                    const image = items[3] === "Up" ? "container-on.png" : "container-off.png";
-                    containers.push(new DockerContainer(
-                        items[0],
-                        items[1],
-                        items[2],
-                        this.context.asAbsolutePath(path.join("resources", image)),
-                        {
-                            command: "docker-explorer.getContainer",
-                            title: "",
-                            arguments: [items[0]],
-                        },
-                    ));
-                }
+            this.containerStrings = Executor.execSync("docker ps -a --format \"{{.ID}} {{.Names}} {{.Image}} {{.Status}}\"")
+                .split(/[\r\n]+/g).filter((item) => item);
+            this.containerStrings.forEach((containerString) => {
+                const items = containerString.split(" ");
+                const image = items[3] === "Up" ? "container-on.png" : "container-off.png";
+                containers.push(new DockerContainer(
+                    items[0],
+                    items[1],
+                    items[2],
+                    this.context.asAbsolutePath(path.join("resources", image)),
+                    {
+                        command: "docker-explorer.getContainer",
+                        title: "",
+                        arguments: [items[1]],
+                    },
+                ));
             });
         } catch (error) {
             if (!this.isErrorMessageShown) {
@@ -54,38 +74,38 @@ export class DockerContainers implements vscode.TreeDataProvider<DockerContainer
         return Promise.resolve(containers);
     }
 
-    public getContainer(containerId: string): void {
-        Executor.runInTerminal(`docker ps -a --filter "id=${containerId}"`);
+    public getContainer(containerName: string): void {
+        Executor.runInTerminal(`docker ps -a --filter "name=${containerName}"`);
         AppInsightsClient.sendEvent("getContainer");
     }
 
-    public startContainer(containerId: string): void {
-        Executor.runInTerminal(`docker start ${containerId}`);
+    public startContainer(containerName: string): void {
+        Executor.runInTerminal(`docker start ${containerName}`);
         AppInsightsClient.sendEvent("startContainer");
     }
 
-    public stopContainer(containerId: string): void {
-        Executor.runInTerminal(`docker stop ${containerId}`);
+    public stopContainer(containerName: string): void {
+        Executor.runInTerminal(`docker stop ${containerName}`);
         AppInsightsClient.sendEvent("stopContainer");
     }
 
-    public restartContainer(containerId: string): void {
-        Executor.runInTerminal(`docker restart ${containerId}`);
+    public restartContainer(containerName: string): void {
+        Executor.runInTerminal(`docker restart ${containerName}`);
         AppInsightsClient.sendEvent("restartContainer");
     }
 
-    public showContainerStatistics(containerId: string): void {
-        Executor.runInTerminal(`docker stats ${containerId}`);
+    public showContainerStatistics(containerName: string): void {
+        Executor.runInTerminal(`docker stats ${containerName}`);
         AppInsightsClient.sendEvent("showContainerStatistics");
     }
 
-    public showContainerLogs(containerId: string): void {
-        Executor.runInTerminal(`docker logs ${containerId}`);
+    public showContainerLogs(containerName: string): void {
+        Executor.runInTerminal(`docker logs ${containerName}`);
         AppInsightsClient.sendEvent("showContainerLogs");
     }
 
-    public removeContainer(containerId: string): void {
-        Executor.runInTerminal(`docker rm ${containerId}`);
+    public removeContainer(containerName: string): void {
+        Executor.runInTerminal(`docker rm ${containerName}`);
         AppInsightsClient.sendEvent("removeContainer");
     }